Refactor component structure and add BrandsCarousel to Homepage
