+++ src/App.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Routes, Route } from 'react-router-dom';
-import Layout from './components/Layout';
+import Layout from './pages/Layout';
 import Homepage from './pages/Homepage/Homepage';
 import About from './pages/About/About';
 import './App.css';
