+++ src/pages/Homepage/Homepage.jsx
@@ -1,11 +1,13 @@
 import React from 'react';
 import './Homepage.css';
 import Jumbotron from './components/Jumbotron';
+import BrandsCarousel from './components/BrandsCarousel';
 
 export default function Homepage() {
     return (
         <>
             <Jumbotron />
+            <BrandsCarousel />
         </>
     );
 }
